fix(data-collector): 修复 SQLite 主键自增和 BaoStock 接口问题
1. models.py: BigInteger 主键改为 Integer，修复 SQLite autoincrement 不生效问题
2. collector.py:
   - get_all_stocks 改用 query_all_stock + 代码前缀过滤，避免 KeyError
   - code 存储格式改为纯6位数字，与 Prisma 查询格式一致
   - query_all_stock 使用昨天日期，避免非交易日返回空列表
3. database.py: 兼容 Prisma 格式的 DATABASE_URL 环境变量

+++ data-collector/models.py
@@ -17,7 +17,7 @@ class StockDaily(Base):
     """股票日K线数据"""
     __tablename__ = "stock_daily"
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     code = Column(String(10), nullable=False, comment="股票代码")
     name = Column(String(50), comment="股票名称")
     trade_date = Column(Date, nullable=False, comment="交易日期")
@@ -46,7 +46,7 @@ class DragonTigerList(Base):
     """龙虎榜数据"""
     __tablename__ = "dragon_tiger_list"
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     trade_date = Column(Date, nullable=False, comment="交易日期")
     code = Column(String(10), nullable=False, comment="股票代码")
     name = Column(String(50), comment="股票名称")
@@ -74,7 +74,7 @@ class DragonTigerDetail(Base):
     """龙虎榜营业部明细"""
     __tablename__ = "dragon_tiger_detail"
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     trade_date = Column(Date, nullable=False, comment="交易日期")
     code = Column(String(10), nullable=False, comment="股票代码")
     name = Column(String(50), comment="股票名称")
@@ -99,7 +99,7 @@ class CollectionLog(Base):
     """数据采集日志"""
     __tablename__ = "collection_log"
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     collection_date = Column(Date, nullable=False, comment="采集日期")
     data_type = Column(String(50), nullable=False, comment="数据类型")
     status = Column(String(20), nullable=False, comment="状态(success/failed)")
